Accept multiple client_id/secret pairs for FxA (bug 1080039)
