+++ sites/identitystage/settings_mkt.py
@@ -132,3 +132,6 @@ FXA_AUTH_DOMAIN = getattr(private_mkt, 'FXA_AUTH_DOMAIN', '')
 FXA_OAUTH_URL = getattr(private_mkt, 'FXA_OAUTH_URL', '')
 FXA_CLIENT_ID = getattr(private_mkt, 'FXA_CLIENT_ID', '')
 FXA_CLIENT_SECRET = getattr(private_mkt, 'FXA_CLIENT_SECRET', '')
+FXA_SECRETS = {
+    FXA_CLIENT_ID: FXA_CLIENT_SECRET,
+}
